% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folder_structure.R
\name{create_folder_structure}
\alias{create_folder_structure}
\title{Check and create folder structure}
\usage{
create_folder_structure(project_name, wd)
}
\arguments{
\item{project_name}{Name of project, for example the name of the forest}

\item{wd}{Working directory where the project is}
}
\value{
create and check folder structure
}
\description{
Check and create folder structure
}
