% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_from_ign.R
\name{download_7z}
\alias{download_7z}
\title{retrieve list of ressources from an url}
\usage{
download_7z(url, src_folder, db_ign)
}
\arguments{
\item{url}{url of the data to download.}

\item{src_folder}{source folder where data are stored.}

\item{db_ign}{character corresponding to the name of the database. Can be one of
\code{"bdalti"}, \code{"bdforet"}, \code{"bdtopo"}, \code{"parcellaire-express"}, \code{"rgealti"}.}
}
\description{
retrieve list of ressources from an url
}
