% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gis_db.R
\name{download_gis_db}
\alias{download_gis_db}
\title{Wrapper around load_from_ign to download all needed database at the right source folder}
\usage{
download_gis_db(num_dep = "75")
}
\arguments{
\item{num_dep}{character corresponding to a department number. "Outre-mer"
departement are not supported for the moment}
}
\value{
all database extract to src_folder
}
\description{
Wrapper around load_from_ign to download all needed database at the right source folder
}
\examples{
\dontrun{
download_gis_db("29")
}
}
