% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_scan.R
\name{download_scan}
\alias{download_scan}
\title{download scan25 and scan100}
\usage{
download_scan(
  shape,
  project_name,
  wd,
  personal_key = "slz53dkjac05e9r547kmzarv"
)
}
\arguments{
\item{shape}{Object of class \code{sf}}

\item{project_name}{Name of project, for example the name of the forest}

\item{wd}{Working directory where the project is}

\item{personal_key}{key from you're personnal IGN webservice account}
}
\value{
.tif
}
\description{
download scan25 and scan100
}
