% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_vector_data.R
\name{download_vector_data}
\alias{download_vector_data}
\title{download_vector_data}
\usage{
download_vector_data(shape, project_name, wd)
}
\arguments{
\item{shape}{Object of class \code{sf}}

\item{project_name}{Name of project, for example the name of the forest}

\item{wd}{Working directory where the project is}
}
\value{
.shp
}
\description{
download_vector_data
}
