% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liste_dep.R
\docType{data}
\name{liste_dep}
\alias{liste_dep}
\title{liste des numero de departements}
\format{
An object of class \code{character} of length 101.
}
\usage{
liste_dep
}
\description{
liste des numero de departements
}
\keyword{datasets}
