% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_from_ign.R
\name{load_from_ign}
\alias{load_from_ign}
\title{Wrapper function to download some IGN database}
\usage{
load_from_ign(db_ign = "bdtopo", num_dep = "75", src_folder = src_folder)
}
\arguments{
\item{db_ign}{character corresponding to the name of the database. Can be one of
\code{"bdalti"}, \code{"bdforet"}, \code{"bdtopo"}, \code{"parcellaire-express"}, \code{"rgealti"}.}

\item{num_dep}{character corresponding to a department number.}

\item{src_folder}{source folder where data are stored.}
}
\value{
All data extract to src_folder
}
\description{
Wrapper function to download some IGN database
}
\examples{
\dontrun{

# create src_folder
my_dir <- file.path(getwd(), "general")
dir.create(my_dir)

# download data
load_from_ign(db_ign = "bdalti",
              num_dep = "75",
              src_folder = my_dir)

# delete src_folder because it is only an example
unlink(basename(my_dir), recursive = TRUE)

}
}
