% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{shape_test}
\alias{shape_test}
\title{test data which is used internally in the package}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
shape_test
}
\description{
test data which is used internally in the package
}
\keyword{datasets}
