% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/situation_map.R
\name{situation_map}
\alias{situation_map}
\title{Create situation map}
\usage{
situation_map(shape, project_name, wd, scan = "scan25")
}
\arguments{
\item{shape}{Object of class \code{sf}}

\item{project_name}{Name of project, for example the name of the forest}

\item{wd}{Working directory where the project is}

\item{scan}{Can be "scan25" or "scan100"}
}
\value{
map and create pdf
}
\description{
Create situation map
}
